Use `10.0` instead of `10.toDouble()`

+++ lib/src/scale/util/nice_range.dart
@@ -20,9 +20,9 @@ num _tickIncrement(
                 : error >= _e2
                     ? 2
                     : 1) *
-        dart_math.pow(10.toDouble(), power);
+        dart_math.pow(10.0, power);
   }
-  return -dart_math.pow(10.toDouble(), -power) /
+  return -dart_math.pow(10.0, -power) /
       (error >= _e10
           ? 10
           : error >= _e5
